requirements.txt
django==4.1
djangorestframework==3.14.0
django-guardian==2.4.0